Fragment廃止へ (LogListFragment削除 step1)
