+++ app/build.gradle
@@ -107,10 +107,11 @@ dependencies {
     androidTestImplementation(composeBom)
 
     implementation "androidx.compose.runtime:runtime"
+    implementation 'androidx.activity:activity-compose:1.7.2'
     implementation "androidx.compose.ui:ui"
     implementation "androidx.compose.foundation:foundation"
     implementation "androidx.compose.foundation:foundation-layout"
-    implementation "androidx.compose.material3:material3:1.1.0"
+    implementation "androidx.compose.material3:material3:1.1.1"
     implementation "androidx.compose.runtime:runtime-livedata"
     implementation "androidx.compose.ui:ui-tooling"
     implementation "com.google.accompanist:accompanist-themeadapter-material:0.28.0"
