+++ app/src/main/java/org/tirasweel/drivelogger/fragments/LogListFragment.kt
@@ -5,30 +5,13 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Toast
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
-import io.realm.kotlin.ext.query
-import io.realm.kotlin.query.Sort
-import kotlinx.serialization.encodeToString
-import kotlinx.serialization.json.Json
 import org.tirasweel.drivelogger.BuildConfig
-import org.tirasweel.drivelogger.R
-import org.tirasweel.drivelogger.compose.DriveLogListScreen
-import org.tirasweel.drivelogger.compose.DriveLogListTopAppBarClickListener
 import org.tirasweel.drivelogger.db.DriveLog
-import org.tirasweel.drivelogger.ui.theme.DriveLoggerTheme
-import org.tirasweel.drivelogger.utils.ConfirmDialogFragment
-import org.tirasweel.drivelogger.utils.RealmUtil
 import org.tirasweel.drivelogger.viewmodels.DriveLogListViewModel
-import java.io.File
-import java.io.FileWriter
 
 class LogListFragment : Fragment() {
 
@@ -42,31 +25,6 @@ class LogListFragment : Fragment() {
     }
 
     private val viewModel: DriveLogListViewModel by viewModels()
-
-    /**
-     * @brief リストのソート順
-     */
-    enum class SortOrderType {
-        AscendingDate,
-        DescendingDate;
-
-        /**
-         * @brief ソート順(RealmのSortで)
-         */
-        val order
-            get() = when (this) {
-                AscendingDate -> Sort.ASCENDING
-                DescendingDate -> Sort.DESCENDING
-            }
-
-        /**
-         * @brief ソートに使用するプロパティ
-         */
-        val property
-            get() = when (this) {
-                AscendingDate, DescendingDate -> "date"
-            }
-    }
 
     private var listener: LogListInteractionListener? = null
 
@@ -87,22 +45,6 @@ class LogListFragment : Fragment() {
             // is destroyed
             setViewCompositionStrategy(ViewCompositionStrategy.DisposeOnViewTreeLifecycleDestroyed)
             setContent {
-                DriveLoggerTheme {
-                    Surface(
-                        modifier = Modifier.fillMaxSize(),
-                        color = MaterialTheme.colorScheme.background
-                    ) {
-                        DriveLogListScreen(
-                            driveLogListViewModel = viewModel,
-                            clickListener = listener,
-                            appBarClickListener = object : DriveLogListTopAppBarClickListener {
-                                override fun onClickExport() {
-                                    executeExport()
-                                }
-                            },
-                        )
-                    }
-                }
             }
         }
     }
@@ -110,47 +52,6 @@ class LogListFragment : Fragment() {
     override fun onResume() {
         super.onResume()
         viewModel.reloadDriveLogs()
-    }
-
-    private fun executeExport() {
-        context?.getExternalFilesDir("DriveLogs")?.let { dir ->
-            val file = File(dir, "export.json")
-
-            val export = {
-                val writer = FileWriter(file)
-
-                val realm = RealmUtil.createRealm()
-                val driveLogs = realm.query<DriveLog>().find()
-
-                driveLogs.forEach { log ->
-                    writer.write(Json.encodeToString(log))
-                }
-
-                writer.close()
-
-                Toast.makeText(context, R.string.message_export_file_successful, Toast.LENGTH_LONG)
-                    .show()
-            }
-
-            // 既にファイルが存在する場合、ダイアログで確認する.
-            if (file.exists()) {
-                val dialog = ConfirmDialogFragment.newInstance(
-                    this,
-                    null,
-                    getString(R.string.message_export_file_already_exists)
-                ) { response ->
-                    if (response) {
-                        export()
-                    } else {
-                        return@newInstance
-                    }
-                }
-                dialog.show(childFragmentManager, "OVERWRITE_CHECK_EXPORT_FILE")
-            } else {
-                export()
-            }
-
-        }
     }
 
     override fun onAttach(context: Context) {
