+++ app/src/main/java/org/tirasweel/drivelogger/utils/ConfirmDialogFragment.kt
@@ -1,5 +1,6 @@
 package org.tirasweel.drivelogger.utils
 
+import android.app.Activity
 import android.app.AlertDialog
 import android.app.Dialog
 import android.os.Bundle
@@ -27,6 +28,27 @@ class ConfirmDialogFragment : DialogFragment() {
     }
 
     companion object {
+        fun createAlertDialog(
+            activity: Activity,
+            title: String?,
+            message: String,
+            listener: (Boolean) -> Unit
+        ): AlertDialog {
+
+            val dialog = AlertDialog.Builder(activity)
+                .setTitle(title)
+                .setMessage(message)
+                .setPositiveButton(R.string.yes) { _, _ ->
+                    listener(true)
+                }
+                .setNegativeButton(R.string.no) { _, _ ->
+                    listener(false)
+                }
+                .create()
+
+            return dialog
+        }
+
         fun newInstance(
             fragment: Fragment,
             title: String?,
